// client/src/app/core/models/creature.model.ts
export interface CreatureTemplateDto {
  id?: number;
  name: string;
  description?: string;
  size?: string;
  type?: string;
  alignment?: string;
  armorClass?: number;
  hitPoints?: number;
  speed?: string;
  strength?: number;
  dexterity?: number;
  constitution?: number;
  intelligence?: number;
  wisdom?: number;
  charisma?: number;
  challengeRating?: number;
}